+++ app/helpers/broadcast_helper.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-# Pulls first image tag from description if one exists.
-module BroadcastHelper
-  def first_image(broadcast)
-    match = simple_markdown(broadcast.description).match(/<img.*?>/m)
-    match[0].to_s.html_safe if match
-  end
-end